--- a/lib/commands/test.js
+++ b/lib/commands/test.js
@@ -10,8 +10,6 @@ const fs = require('fs');
 const temp = require('temp');
 temp.track();
 
-require('express').static.mime.define({ 'application/wasm': ['wasm'] });
-
 let defaultPort = 7357;
 
 module.exports = Command.extend({
@@ -80,6 +78,9 @@ module.exports = Command.extend({
 
   init() {
     this._super.apply(this, arguments);
+
+    // Make sure Testem supports the Wasm MIME type:
+    require('express').static.mime.define({ 'application/wasm': ['wasm'] });
 
     this.Builder = this.Builder || Builder;
     this.Watcher = this.Watcher || Watcher;