--- a/lib/models/blueprint.js
+++ b/lib/models/blueprint.js
@@ -23,7 +23,6 @@ const EOL = require('os').EOL;
 const bowEpParser = require('bower-endpoint-parser');
 const logger = require('heimdalljs-logger')('ember-cli:blueprint');
 const normalizeEntityName = require('ember-cli-normalize-entity-name');
-const { removeTypes } = require('remove-types');
 const isAddon = require('../utilities/is-addon');
 const { deprecate } = require('../debug');
 
@@ -532,6 +531,7 @@ let Blueprint = CoreObject.extend({
   async convertToJS(fileInfo) {
     let rendered = await fileInfo.render();
 
+    const { removeTypes } = require('remove-types');
     const transformed = await removeTypes(rendered);
 
     fileInfo.rendered = transformed;