--- a/lib/tasks/interactive-new.js
+++ b/lib/tasks/interactive-new.js
@@ -1,9 +1,5 @@
 'use strict';
 
-const inquirer = require('inquirer');
-const { isLangCode } = require('is-language-code');
-const osLocale = require('os-locale');
-
 const Task = require('../models/task');
 const isValidProjectName = require('../utilities/valid-project-name');
 
@@ -11,6 +7,8 @@ const DEFAULT_LOCALE = 'en-US';
 
 class InteractiveNewTask extends Task {
   async run(newCommandOptions, _testAnswers) {
+    const inquirer = require('inquirer');
+
     let prompt = inquirer.createPromptModule();
     let questions = await this.getQuestions(newCommandOptions);
     let answers = await prompt(questions, _testAnswers);
@@ -24,6 +22,8 @@ class InteractiveNewTask extends Task {
   }
 
   async getQuestions(newCommandOptions = {}) {
+    const { isLangCode } = require('is-language-code');
+
     return [
       {
         name: 'blueprint',
@@ -145,6 +145,8 @@ class InteractiveNewTask extends Task {
   }
 
   getUserLocale() {
+    const osLocale = require('os-locale');
+
     return osLocale();
   }
 }