--- a/lib/tasks/transforms/amd/index.js
+++ b/lib/tasks/transforms/amd/index.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const shimAmd = require('./amd-shim');
-
 class AmdTransformAddon {
   /**
    * This addon is used to register a custom AMD transform for app and addons to use.
@@ -15,6 +13,8 @@ class AmdTransformAddon {
   }
 
   importTransforms() {
+    const shimAmd = require('./amd-shim');
+
     return {
       amd: {
         transform: (tree, options) => {