Quick results:

| Command       | Before        | After         |
|---------------|---------------|---------------|
| ember help    | 400ms - 410ms | 230ms - 240ms |
| ember version | 500ms - 520ms | 250ms - 270ms |

Closes #10171.